refactor: Refactor CSL event handler
Extract the code that retrieves/signs CSLs into a separate service.
