+++ pkg/service/credentialstatus/eventhandler/eventhandler_service.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-//go:generate mockgen -destination service_mocks_test.go -self_package mocks -package eventhandler -source=eventhandler_service.go -mock_names profileService=MockProfileService,kmsRegistry=MockKMSRegistry
+//go:generate mockgen -destination service_mocks_test.go -self_package mocks -package eventhandler -source=eventhandler_service.go -mock_names CSLService=MockCSLService
 
 package eventhandler
 
@@ -13,70 +13,37 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/piprate/json-gold/ld"
 	"github.com/trustbloc/logutil-go/pkg/log"
-	"github.com/trustbloc/vc-go/dataintegrity/models"
 	"github.com/trustbloc/vc-go/verifiable"
-	vcsverifiable "github.com/trustbloc/vcs/pkg/doc/verifiable"
 
 	"github.com/trustbloc/vcs/internal/logfields"
 	"github.com/trustbloc/vcs/pkg/doc/vc"
-	vccrypto "github.com/trustbloc/vcs/pkg/doc/vc/crypto"
 	"github.com/trustbloc/vcs/pkg/doc/vc/statustype"
 	"github.com/trustbloc/vcs/pkg/event/spi"
-	vcskms "github.com/trustbloc/vcs/pkg/kms"
-	profileapi "github.com/trustbloc/vcs/pkg/profile"
 	"github.com/trustbloc/vcs/pkg/service/credentialstatus"
-)
-
-const (
-	defaultRepresentation = "jws"
-
-	jsonKeyProofValue         = "proofValue"
-	jsonKeyProofPurpose       = "proofPurpose"
-	jsonKeyVerificationMethod = "verificationMethod"
-	jsonKeySignatureOfType    = "type"
-	jsonStatusListType        = "type"
 )
 
 var logger = log.New("credentialstatus-eventhandler")
 
-type profileService interface {
-	GetProfile(profileID profileapi.ID, profileVersion profileapi.Version) (*profileapi.Issuer, error)
-}
+const jsonStatusListType = "type"
 
-type kmsRegistry interface {
-	GetKeyManager(config *vcskms.Config) (vcskms.VCSKeyManager, error)
-}
-
-type vcCrypto interface {
-	SignCredential(signerData *vc.Signer, vc *verifiable.Credential,
-		opts ...vccrypto.SigningOpts) (*verifiable.Credential, error)
+type CSLService interface {
+	SignCSL(profileID, profileVersion string, csl *verifiable.Credential) ([]byte, error)
+	GetCSLVCWrapper(ctx context.Context, cslURL string) (*credentialstatus.CSLVCWrapper, error)
+	UpsertCSLVCWrapper(ctx context.Context, cslURL string, wrapper *credentialstatus.CSLVCWrapper) error
 }
 
 type Config struct {
-	CSLVCStore     credentialstatus.CSLVCStore
-	ProfileService profileService
-	KMSRegistry    kmsRegistry
-	Crypto         vcCrypto
-	DocumentLoader ld.DocumentLoader
+	CSLService CSLService
 }
 
 type Service struct {
-	cslStore       credentialstatus.CSLVCStore
-	profileService profileService
-	kmsRegistry    kmsRegistry
-	crypto         vcCrypto
-	documentLoader ld.DocumentLoader
+	cslService CSLService
 }
 
 func New(conf *Config) *Service {
 	return &Service{
-		cslStore:       conf.CSLVCStore,
-		profileService: conf.ProfileService,
-		kmsRegistry:    conf.KMSRegistry,
-		crypto:         conf.Crypto,
-		documentLoader: conf.DocumentLoader,
+		cslService: conf.CSLService,
 	}
 }
 
@@ -90,7 +57,12 @@ func (s *Service) HandleEvent(ctx context.Context, event *spi.Event) error { //n
 
 	payload := credentialstatus.UpdateCredentialStatusEventPayload{}
 
-	jsonData, err := json.Marshal(event.Data.(map[string]interface{}))
+	doc, ok := event.Data.(map[string]interface{})
+	if !ok {
+		return fmt.Errorf("invalid event data")
+	}
+
+	jsonData, err := json.Marshal(doc)
 	if err != nil {
 		return err
 	}
@@ -104,7 +76,7 @@ func (s *Service) HandleEvent(ctx context.Context, event *spi.Event) error { //n
 
 func (s *Service) handleEventPayload(
 	ctx context.Context, payload credentialstatus.UpdateCredentialStatusEventPayload) error {
-	clsWrapper, err := s.getCSLVCWrapper(ctx, payload.CSLURL)
+	clsWrapper, err := s.cslService.GetCSLVCWrapper(ctx, payload.CSLURL)
 	if err != nil {
 		return fmt.Errorf("get CSL VC wrapper failed: %w", err)
 	}
@@ -126,7 +98,7 @@ func (s *Service) handleEventPayload(
 		return fmt.Errorf("failed to update status: %w", err)
 	}
 
-	signedCredentialBytes, err := s.signCSL(payload.ProfileID, payload.ProfileVersion, clsWrapper.VC)
+	signedCredentialBytes, err := s.cslService.SignCSL(payload.ProfileID, payload.ProfileVersion, clsWrapper.VC)
 	if err != nil {
 		return fmt.Errorf("failed to sign CSL: %w", err)
 	}
@@ -135,119 +107,11 @@ func (s *Service) handleEventPayload(
 		VCByte: signedCredentialBytes,
 	}
 
-	if err = s.cslStore.Upsert(ctx, payload.CSLURL, vcWrapper); err != nil {
-		return fmt.Errorf("cslStore.Upsert failed: %w", err)
+	if err = s.cslService.UpsertCSLVCWrapper(ctx, payload.CSLURL, vcWrapper); err != nil {
+		return fmt.Errorf("save CSL failed: %w", err)
 	}
 
 	return nil
-}
-
-func (s *Service) signCSL(profileID, profileVersion string, csl *verifiable.Credential) ([]byte, error) {
-	issuerProfile, err := s.profileService.GetProfile(profileID, profileVersion)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get profile: %w", err)
-	}
-
-	keyManager, err := s.kmsRegistry.GetKeyManager(issuerProfile.KMSConfig)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get KMS: %w", err)
-	}
-
-	signer := &vc.Signer{
-		Format:                  issuerProfile.VCConfig.Format,
-		DID:                     issuerProfile.SigningDID.DID,
-		Creator:                 issuerProfile.SigningDID.Creator,
-		KMSKeyID:                issuerProfile.SigningDID.KMSKeyID,
-		SignatureType:           issuerProfile.VCConfig.SigningAlgorithm,
-		KeyType:                 issuerProfile.VCConfig.KeyType,
-		KMS:                     keyManager,
-		SignatureRepresentation: issuerProfile.VCConfig.SignatureRepresentation,
-		VCStatusListType:        issuerProfile.VCConfig.Status.Type,
-		SDJWT:                   vc.SDJWT{Enable: false},
-		DataIntegrityProof:      issuerProfile.VCConfig.DataIntegrityProof,
-	}
-
-	signOpts, err := prepareSigningOpts(signer, csl.Proofs())
-	if err != nil {
-		return nil, fmt.Errorf("prepareSigningOpts failed: %w", err)
-	}
-
-	signedCredential, err := s.crypto.SignCredential(signer, csl, signOpts...)
-	if err != nil {
-		return nil, fmt.Errorf("sign CSL failed: %w", err)
-	}
-
-	return signedCredential.MarshalJSON()
-}
-
-func (s *Service) getCSLVCWrapper(ctx context.Context, cslURL string) (*credentialstatus.CSLVCWrapper, error) {
-	vcWrapper, err := s.cslStore.Get(ctx, cslURL)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get CSL from store: %w", err)
-	}
-
-	cslVC, err := verifiable.ParseCredential(vcWrapper.VCByte,
-		verifiable.WithDisabledProofCheck(),
-		verifiable.WithJSONLDDocumentLoader(s.documentLoader))
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse CSL: %w", err)
-	}
-
-	vcWrapper.VC = cslVC
-
-	return vcWrapper, nil
-}
-
-// prepareSigningOpts prepares signing opts from recently issued proof of given credential.
-func prepareSigningOpts(profile *vc.Signer, proofs []verifiable.Proof) ([]vccrypto.SigningOpts, error) {
-	var signingOpts []vccrypto.SigningOpts
-
-	if len(proofs) == 0 {
-		return signingOpts, nil
-	}
-
-	// pick latest proof if there are multiple
-	proof := proofs[len(proofs)-1]
-
-	representation := defaultRepresentation
-	if _, ok := proof[jsonKeyProofValue]; ok {
-		representation = jsonKeyProofValue
-	}
-
-	signingOpts = append(signingOpts, vccrypto.WithSigningRepresentation(representation))
-
-	purpose, err := getStringValue(jsonKeyProofPurpose, proof)
-	if err != nil {
-		return nil, err
-	}
-
-	signingOpts = append(signingOpts, vccrypto.WithPurpose(purpose))
-
-	vm, err := getStringValue(jsonKeyVerificationMethod, proof)
-	if err != nil {
-		return nil, err
-	}
-
-	// add verification method option only when it is not matching profile creator
-	if vm != profile.Creator {
-		signingOpts = append(signingOpts, vccrypto.WithVerificationMethod(vm))
-	}
-
-	signTypeName, err := getStringValue(jsonKeySignatureOfType, proof)
-	if err != nil {
-		return nil, err
-	}
-
-	if signTypeName != "" && signTypeName != models.DataIntegrityProof {
-		signType, err := vcsverifiable.GetSignatureTypeByName(signTypeName)
-		if err != nil {
-			return nil, err
-		}
-
-		signingOpts = append(signingOpts, vccrypto.WithSignatureType(signType))
-	}
-
-	return signingOpts, nil
 }
 
 func getStringValue(key string, vMap map[string]interface{}) (string, error) {
